+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Fixed a bug where RSS always used current site instead of requested site
 - Fixed a bug where only podcast for the default site was displayed on the permission list
 - Fixed defineSources for episode for multi-site
+- Fixed a bug where it was possible to create episodes for sites that podcast was not propagated to
 - Removed getPodcastBySlug()
 - Dropped Craft 4.3 support